+++ redwood-flex-container/src/commonMain/kotlin/app/cash/redwood/flexcontainer/Measurable.kt
@@ -22,7 +22,7 @@ public open class Measurable {
    * The attribute is about how wide the view wants to be. Can be one of the
    * constants [MatchParent] or [WrapContent], or an exact size.
    */
-  public open val requestedWidth: Int = WrapContent
+  public open val requestedWidth: Double = WrapContent
 
   /**
    * The requested height of the item.
@@ -30,50 +30,50 @@ public open class Measurable {
    * The attribute is about how wide the view wants to be. Can be one of the
    * constants [MatchParent] or [WrapContent], or an exact size.
    */
-  public open val requestedHeight: Int = WrapContent
+  public open val requestedHeight: Double = WrapContent
 
   /**
    * The minimum width attribute of the item.
    *
    * The attribute determines the minimum width the child can shrink to.
    */
-  public open val minWidth: Int = 0
+  public open val minWidth: Double = 0.0
 
   /**
    * The minimum height attribute of the item.
    *
    * The attribute determines the minimum height the child can shrink to.
    */
-  public open val minHeight: Int = 0
+  public open val minHeight: Double = 0.0
 
   /**
    * The maximum width attribute of the item.
    *
    * The attribute determines the maximum width the child can expand to.
    */
-  public open val maxWidth: Int = Int.MAX_VALUE
+  public open val maxWidth: Double = Double.MAX_VALUE
 
   /**
    * The maximum height attribute of the item.
    *
    * The attribute determines the maximum height the child can expand to.
    */
-  public open val maxHeight: Int = Int.MAX_VALUE
+  public open val maxHeight: Double = Double.MAX_VALUE
 
   /**
    * Return the item's width given a fixed [height].
    */
-  public open fun width(height: Int): Int = measure(
-    widthSpec = MeasureSpec.from(0, MeasureSpecMode.Unspecified),
+  public open fun width(height: Double): Double = measure(
+    widthSpec = MeasureSpec.from(0.0, MeasureSpecMode.Unspecified),
     heightSpec = MeasureSpec.from(height, MeasureSpecMode.Exactly),
   ).width
 
   /**
    * Return the item's height given a fixed [width].
    */
-  public open fun height(width: Int): Int = measure(
+  public open fun height(width: Double): Double = measure(
     widthSpec = MeasureSpec.from(width, MeasureSpecMode.Exactly),
-    heightSpec = MeasureSpec.from(0, MeasureSpecMode.Unspecified),
+    heightSpec = MeasureSpec.from(0.0, MeasureSpecMode.Unspecified),
   ).height
 
   public open fun measure(widthSpec: MeasureSpec, heightSpec: MeasureSpec): Size {
@@ -88,12 +88,12 @@ public open class Measurable {
      * A special constant for [requestedWidth] or [requestedHeight] that means that the item wants to be as big as its
      * parent.
      */
-    public const val MatchParent: Int = -1
+    public const val MatchParent: Double = -1.0
 
     /**
      * A special constant for [requestedWidth] or [requestedHeight] that the item wants to be just large enough to fit
      * its own internal content, taking its own padding into account.
      */
-    public const val WrapContent: Int = -2
+    public const val WrapContent: Double = -2.0
   }
 }
